requirements.txt
lightning-utilities==0.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
MedPy==0.4.0
ml-collections==0.1.0
numpy==1.20.3
opencv-python==4.5.5.62
opencv-python-headless==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
pytorch-fid==0.2.0
pytorch-lightning==1.8.5.post0
pytorch-training==0.0.1
PyYAML==5.4.1
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.6.3
statsmodels==0.12.2
# Editable install with no version control (synthesis-in-style-lightning==0.0.1)
-e /workspace/final_application/synthesis_in_style_lightning/stylegan_code_finder
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
torch==1.9.0+cu111
torchmetrics==0.11.0
torchvision==0.10.0+cu111
tqdm==4.62.2
wandb==0.12.9
